+++ frontend/src/js/elements/scroll-to-top.js
@@ -1,4 +1,4 @@
-import * as ActivePage from "./../elements/active-page";
+import * as ActivePage from "./../states/active-page";
 
 let visible = false;
